requirements.txt
aiosmtpd
python-dotenv